Add tmate config
tmate uses an older version of tmux, so the normal tmux file doesn't
load properly.

This commit adds a tmate configuration that matches the normal tmux
file, but uses an older syntax.

When using tmate, tmux errors will display because the tmux config file
will be loaded as well. Despite this, it appears to work as expected.